% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudcheck.R
\name{azcheck}
\alias{azcheck}
\title{Tests that azure libraries are installed and working.}
\usage{
azcheck()
}
\value{
T/F depending on if az is setup and works
}
\description{
Tests that azure libraries are installed and working.
}
