% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/azure_setup.R
\name{azure_check}
\alias{azure_check}
\title{azure_setup.R
DRAFT
functions to simplify set up Azure options for the azuresub and subscription
and other Azure values that should not go into this library
Tests that azure libraries are installed and working.}
\usage{
azure_check()
}
\value{
T/F depending on if az is setup and works
}
\description{
azure_setup.R
DRAFT
functions to simplify set up Azure options for the azuresub and subscription
and other Azure values that should not go into this library
Tests that azure libraries are installed and working.
}
