% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/azure_vm.R
\name{deleteDeploymentResources}
\alias{deleteDeploymentResources}
\title{Delete all the resources created by a deployment of a vm
This will use a tag created by a deployment to delete all resources
Deletion may take a minute or two for all resources to be cleaned up}
\usage{
deleteDeploymentResources(vmName, rgName = getOption("azurerg"))
}
\arguments{
\item{vmName}{the name of the vm created by the deployment to be deleted}

\item{rgName}{the resource group of the resources, default to the get option azurerg}
}
\description{
Delete all the resources created by a deployment of a vm
This will use a tag created by a deployment to delete all resources
Deletion may take a minute or two for all resources to be cleaned up
}
