% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/azure_setup.R
\name{get_azure_login}
\alias{get_azure_login}
\title{access an existing login to azure or create a new one
if a new login is to be created, there will be a link pasted in the console
that must be visited, instructions are displayed in the console as well}
\usage{
get_azure_login()
}
\value{
the azure login
}
\description{
access an existing login to azure or create a new one
if a new login is to be created, there will be a link pasted in the console
that must be visited, instructions are displayed in the console as well
}
