% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/azure_setup.R
\name{get_container}
\alias{get_container}
\title{get the container object}
\usage{
get_container(
  azurecontainer = getOption("azurecontainer"),
  azurestor = getOption("azurestor"),
  rgname = getOption("azurerg"),
  storageaccesskey = getOption("storageaccesskey")
)
}
\arguments{
\item{azurecontainer}{the name of the container, defaults to the getoption}

\item{azurestor}{the name of the storage account, defaults to the getoption}

\item{rgname}{the name of the resource group, defaults to the getoption}

\item{storageaccesskey}{the storage access key for the given storage account, defaults to the getoption}
}
\value{
the container object, null if it doesn't exist
}
\description{
get the container object
}
