% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/azure_setup.R
\name{get_stor}
\alias{get_stor}
\title{get the storage account object}
\usage{
get_stor(azurestor = getOption("azurestor"), rgname = getOption("azurerg"))
}
\arguments{
\item{azurestor}{the name of the storage account, defaults to the getoption}

\item{rgname}{the name of the resource group, defaults to the getoption}
}
\value{
the storage account object, null if it doesn't exist
}
\description{
get the storage account object
}
