% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/azure_setup.R
\name{get_sub}
\alias{get_sub}
\title{get the azure subscription object for the given sub id value
can also be used to test if a the azuresub is valid (will return NULL)
requires Azure login and this function will initiate that}
\usage{
get_sub(azuresub = getOption("azuresub"))
}
\arguments{
\item{azuresub}{optional string of subscriptoin id e.g xxxxxxxx-xxx-xxx-xxxx-xxxxxxxxxxxx}
}
\value{
AzureRMR subscription object, or NULL if invalid sub id
}
\description{
get the azure subscription object for the given sub id value
can also be used to test if a the azuresub is valid (will return NULL)
requires Azure login and this function will initiate that
}
