% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/azure_vm.R
\name{get_vm}
\alias{get_vm}
\title{get the desired vm, if it exists}
\usage{
get_vm(vmName, rgName = getOption("azurerg"), rg = NULL, verbose = FALSE)
}
\arguments{
\item{vmName}{the name of the desired vm}

\item{rgName}{the resource group of the vm, default to the get option azurerg}

\item{rg}{the resource group object, used to make the function call less demanding if already known, default to NULL}
}
\value{
the vm if it exists, else null
}
\description{
get the desired vm, if it exists
}
