% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/azure_setup.R
\name{new_rg}
\alias{new_rg}
\title{Create a new resource group, set it to the azurerg option}
\usage{
new_rg(azurerg = getOption("azurerg"), azuresub = getOption("azuresub"))
}
\arguments{
\item{azurerg}{the name of the resource group to create}

\item{azuresub}{the subscription id to create the resource group in}
}
\value{
the resource group object of either the newly created resource group, or the resource group that already exists.
}
\description{
Create a new resource group, set it to the azurerg option
}
