% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/azure_setup.R
\name{set_storage_account}
\alias{set_storage_account}
\title{set options for the azure storage account to use for this session}
\usage{
set_storage_account()
}
\value{
T/F if valid values were sent
}
\description{
set options for the azure storage account to use for this session
}
