% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/azure_setup.R
\name{set_storage_options}
\alias{set_storage_options}
\title{set options for the azure storage account to use for this session
These parameters can either be explicitly set using this function
otherwise, the values within the .Renviron file will be used}
\usage{
set_storage_options(
  azurestor = NULL,
  azurecontainer = NULL,
  storageaccesskey = NULL
)
}
\arguments{
\item{azurestor}{the name of the storage account to be used}

\item{azurecontainer}{the name of the storage container to be used}

\item{storageaccesskey}{the storage access key for the storage account to be used, found in the portal under "Access keys" in the storage account menu}
}
\value{
T/F if valid values were sent
}
\description{
set options for the azure storage account to use for this session
These parameters can either be explicitly set using this function
otherwise, the values within the .Renviron file will be used
}
