% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/azure_vm.R
\name{vm_from_template}
\alias{vm_from_template}
\title{Launch a vm from a template file and a extension file provided
Remember to set your working directory prior to use.}
\usage{
vm_from_template(
  vmName,
  templateFile,
  shellScript,
  adminPasswordOrKey,
  userPassword,
  cpuSize = c("CPU-4GB", "CPU-7GB", "CPU-8GB", "CPU-14GB", "CPU-16GB", "GPU-56GB"),
  ubuntuOSVersion = c("18.04-LTS", "20_04-lts", "20_04-daily-lts-gen2"),
  adminUsername = getOption("azureuser"),
  webUsername = getOption("azureuser"),
  dnsNameForPublicIP = getOption("azureuser"),
  storageAccount = getOption("azurestor"),
  scriptContainer = getOption("azurecontainer"),
  storageKey = getOption("storageaccesskey"),
  resourceGroup = getOption("azurerg"),
  storageContainer = getOption("azurecontainer"),
  wait = TRUE
)
}
\arguments{
\item{vmName}{the name of the VM, also used as a prefix on all related resources created during the deployment}

\item{templateFile}{the file path to the template json used to deploy the VM and other resources}

\item{shellScript}{the file path to the extension script file}

\item{adminPasswordOrKey}{ssh public key used to access the vm through ssh}

\item{userPassword}{Rstudio password}

\item{cpuSize}{the size of the cpu, one of the following list ("CPU-4GB", "CPU-7GB", "CPU-8GB", "CPU-14GB", "CPU-16GB", "GPU-56GB")}

\item{ubuntuOSVersion}{the Ubuntu version of the VM, one of the following list ("18.04-LTS", "20_04-lts", "20_04-daily-lts-gen2")}

\item{adminUsername}{the Username used to login to the VM, default to the azureuser option}

\item{webUsername}{the Username used to login to RStudio, default to the azureuser option}

\item{dnsNameForPublicIP}{unique naming for the PublicIP resource, default to the azureuser option}

\item{storageAccount}{the name of the storage account linked to the vm, default to azurestor option}

\item{scriptContainer}{the name of the container where the extension script is hosted, default to azurecontainer option}

\item{storageKey}{the access key to the storage account linked to the vm, default to storageaccesskey option}

\item{resourceGroup}{the name of the resource group that will contain all resources created, default to azurerg option}

\item{storageContainer}{the name of the container to be mounted to the vm}

\item{wait}{if true, wait until the vm deployment is complete to return, default true}
}
\description{
Launch a vm from a template file and a extension file provided
Remember to set your working directory prior to use.
}
