% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/azure_setup.R
\name{set_azure_options}
\alias{set_azure_options}
\title{read azure defaults from the environment to set package options.}
\usage{
set_azure_options(
  azuresub = NULL,
  azurerg = NULL,
  azureuser = NULL,
  azurestor = NULL,
  azurecontainer = NULL,
  storageaccesskey = NULL,
  verbose = FALSE
)
}
\arguments{
\item{azuresub}{the azure subscription ID of the user}

\item{azurerg}{the name of the resource group to be used}

\item{azureuser}{the azure username of the user}

\item{azurestor}{the name of the storage account to be used}

\item{azurecontainer}{the name of the storage container to be used}

\item{storageaccesskey}{the storage access key for the storage account to be used, found in the portal under "Access keys" in the storage account menu}

\item{verbose}{TRUE/FALSE run in verbose mode.}
}
\value{
T/F depending on if az is setup and works.
}
\description{
call this function prior to using any of the functions in this library
this perhaps should be called when the package loads and/or at beginning of most of the functions
usage
  stopifnot(set_azure_options())
 ... your code to work with azure
}
\details{
The parameters can be explicitly set using this function
otherwise the values set in the .Renviron file
will be used.

If the values in Renviron, or sent as params are not valid, the options are not set
}
